pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "twine", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "bielik"
version = "0.2.8"  # Major version bump for Conda transition
description = "Bielik — Local-first AI assistant with HuggingFace and SpeakLeash models"
authors = [{name = "Tom Sapletta"}]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.9"  # Bump minimum Python version for better compatibility
keywords = ["llm", "chat", "huggingface", "bielik", "polish", "ai", "assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Natural Language :: Polish",
    "Natural Language :: English"
]

# Core Dependencies - Minimal set required for basic functionality
dependencies = [
    # Core functionality
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    
    # File and document processing
    "pypdf>=3.0.0,<4.0.0",
    "python-docx>=0.8.11,<1.0.0",
    "beautifulsoup4>=4.8.0,<5.0.0",
    "html2text>=2020.1.16,<2025.0.0",
    
    # HuggingFace integration (core)
    "huggingface-hub>=0.16.0,<1.0.0",
    "tokenizers>=0.13.0,<1.0.0",
    "sentencepiece>=0.1.99,<1.0.0",
    "protobuf>=4.21.0,<5.0.0"
]

[project.optional-dependencies]
# Local model support (CPU/GPU)
local = [
    "llama-cpp-python>=0.2.0,<1.0.0",
    "sentence-transformers>=2.2.0,<3.0.0"
]

# GPU acceleration (includes local)
gpu = [
    "bielik[local]",
    "torch>=2.0.0,<3.0.0",
    "transformers>=4.30.0,<5.0.0",
    "accelerate>=0.20.0,<1.0.0",
    "bitsandbytes>=0.40.0,<1.0.0"
]

# Full AI capabilities (includes gpu)
ai = [
    "bielik[gpu]",
    "transformers[torch]>=4.30.0,<5.0.0",
    "datasets>=2.12.0,<3.0.0",
    "evaluate>=0.4.0,<1.0.0"
]

# Development tools
dev = [
    "bielik[ai]",  # Include all AI capabilities
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "flake8>=6.0.0,<7.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "mypy>=1.0.0,<2.0.0",
    "build>=0.10.0,<1.0.0",
    "twine>=4.0.0,<5.0.0",
    "pre-commit>=3.0.0,<4.0.0"
]

# Documentation
docs = [
    "mkdocs>=1.4.0,<2.0.0",
    "mkdocs-material>=9.0.0,<10.0.0",
    "mkdocstrings[python]>=0.20.0,<1.0.0"
]

# Testing infrastructure
test = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-asyncio>=0.20.0,<1.0.0"
]

# CLI completion
completion = [
    "click-completion>=0.5.0,<1.0.0",
    "shellingham>=1.5.0,<2.0.0"
]

# Web interface
web = [
    "fastapi>=0.95.0,<1.0.0",
    "uvicorn[standard]>=0.21.0,<1.0.0",
    "jinja2>=3.0.0,<4.0.0",
    "python-multipart>=0.0.6,<1.0.0"
]

# Everything (for development and testing)
all = [
    "bielik[ai,dev,docs,test,completion,web]"
]

[project.scripts]
bielik = "bielik.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["bielik"]
include-package-data = true
